+++ spec/directions/move_solutions_to_answer_key/default_strategy_spec.rb
@@ -55,7 +55,7 @@ RSpec.describe Kitchen::Directions::MoveSolutionsToAnswerKey::V1 do
         metadata_source: metadata_element,
         strategy: :default,
         append_to: append_to,
-        strategy_options: { classes: %w[review-questions] }
+        strategy_options: { selectors: %w[section.review-questions] }
       )
     end