+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+* Changed selector expected by `default strategy` in `move_solutions_to_answer_key` to optimize search (major)
 
 * Fix Rubocop GitHub Action's regular expression used to select files to lint (patch)